fix imports and move Filters to one folder
