+++ src/components/Filters/__tests__/FilterPrices.test.jsx
